requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
onnxruntime>=1.15.0
numpy>=1.24.0
piper-tts>=1.2.0
scipy>=1.11.0
python-multipart>=0.0.6
